+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Utils.writeFileSync
 - add Utils.merge, Utils.derive
 - add Binding.as alias for Binding.transform
+- add MprisPlayer.track_album
 
 ## Fixes