+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Utils.writeFileSync
 - add Utils.merge, Utils.derive
 - add Binding.as alias for Binding.transform
+- add MprisPlayer.track_album
 
 ## Fixes